fix imports from deleted env.ts to test-user.ts after mailpit-verification split
